fix: 잘못된 패키지 리펙터링 + merge 추가 작업
