+++ src/main/java/guesthouse/staffrecruitment/dto/response/StaffRecruitmentPostDto.java
@@ -1,4 +1,4 @@
-package staffrecruitment.dto.response;
+package guesthouse.staffrecruitment.dto.response;
 
 import lombok.Builder;
