+++ src/main/java/guesthouse/staffrecruitment/repository/StaffRecruitmentRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StaffRecruitmentRepository extends JpaRepository<StaffRecruitment, Long> {
+public interface StaffRecruitmentRepository extends JpaRepository<StaffRecruitment, Long>, StaffRecruitmentCustomRepository {
+
 }
