+++ src/main/java/guesthouse/staffrecruitment/repository/StaffRecruitmentRepositoryImp.java
@@ -1,4 +1,4 @@
-package staffrecruitment.repository;
+package guesthouse.staffrecruitment.repository;
 
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.dsl.BooleanExpression;
