--- src/main/java/staffrecruitment/repository/StaffRecruitmentRepository.java
@@ -1,9 +1,0 @@
-package staffrecruitment.repository;
-
-import guesthouse.staffrecruitment.domain.model.StaffRecruitment;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StaffRecruitmentRepository extends JpaRepository<StaffRecruitment, Long>, StaffRecruitmentCustomRepository {
-
-}
-
